step2 calculator after feedback
