+++ src/main/java/study2/Operator.java
@@ -8,11 +8,9 @@ public class Operator { // 연산자 종류 check 및  연산하는 메소드
 	static String DIV = "/";
 	
 	public int add(int a, int b) {
-		
 		return a + b;
 	}
 	public int sub(int a, int b) {
-		
 		return a - b;
 	}
 	public int multi(int a, int b) {
